(PC-42039)[BO] feat: show email changes in pro user history
