+++ api/src/pcapi/routes/backoffice/templates/pro_user/get/details/history.html
@@ -1,4 +1,4 @@
 {% from "components/history/view.html" import build_history_view with context %}
-{% call build_history_view(dst, form, actions, false) %}
+{% call build_history_view(dst, form, history, false) %}
   {{ user.firstName }} {{ user.lastName | upper }}
 {% endcall %}
